+++ docs/changes.rst
@@ -48,6 +48,11 @@ Released: not yet
   update-only properties (e.g. several properties in Partitions).
   (Fixed as part of issue #31).
 
+* Fixed the issue that a partition in "terminated" or "paused" status
+  could not be made absent (i.e. deleted). Now, the partition is
+  stopped which should bring it into "stopped" status, and then
+  deleted. (Issue #29).
+
 **Enhancements:**
 
 **Known issues:**